Moved services to seperate namespace
