--- grade-management-new/GradeManagement.Bll/AssignmentLogService.cs
@@ -1,6 +1,0 @@
-﻿namespace GradeManagement.Bll;
-
-public class AssignmentLogService
-{
-
-}
