+++ grade-management-new/GradeManagement.Bll/Services/AssignmentLogService.cs
@@ -1,0 +1,6 @@
+﻿namespace GradeManagement.Bll.Services;
+
+public class AssignmentLogService
+{
+
+}
