+++ grade-management-new/GradeManagement.Bll/Services/DashboardService.cs
@@ -5,7 +5,7 @@ using GradeManagement.Shared.Dtos;
 
 using Microsoft.EntityFrameworkCore;
 
-namespace GradeManagement.Bll;
+namespace GradeManagement.Bll.Services;
 
 public class DashboardService
 {
