+++ grade-management-new/GradeManagement.QueueFunction/Functions/PullRequestStatusChangedEventFunction.cs
@@ -1,4 +1,5 @@
 using GradeManagement.Bll;
+using GradeManagement.Bll.Services;
 using GradeManagement.Shared.Dtos.AssignmentEvents;
 
 using Microsoft.Azure.Functions.Worker;
