+++ grade-management-new/GradeManagement.QueueFunction/Program.cs
@@ -1,4 +1,5 @@
 using GradeManagement.Bll;
+using GradeManagement.Bll.Services;
 
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.EntityFrameworkCore;
