+++ grade-management-new/GradeManagement.Server/Controllers/PullRequestController.cs
@@ -1,4 +1,5 @@
 ﻿using GradeManagement.Bll;
+using GradeManagement.Bll.Services;
 using GradeManagement.Shared.Dtos;
 
 using Microsoft.AspNetCore.Mvc;
